+++ view.js
@@ -1,4 +1,6 @@
 const { ipcRenderer } = require('electron');
+const { listFromSet } = require("./util");
+
 
 ipcRenderer.on('update-view-metadata', function(event, metadata) {
 	//console.log("update metadata");
@@ -18,6 +20,25 @@ ipcRenderer.on('update-view-metadata', function(event, metadata) {
 		document.title = metadata.title;
 });
 
+ipcRenderer.on('update-view-bookmarks', function(event, bookmarks) {
+	//console.log("update-view-bookmarks");
+	let container = document.getElementById("bookmarkContainer");
+	//remove children
+	while (container.firstChild) {
+		container.removeChild(container.firstChild);
+	}
+
+	if (bookmarks.length > 0) {
+		let list = container.appendChild(document.createElement("ul"));
+		for (let b of bookmarks) {
+			let li = document.createElement("li");
+			let aLabel = "Section #" + b.index; //TODO: improve bookmark labels
+			li.appendChild(linkOrText(aLabel, {index: b.index, url: b.url}));
+			list.appendChild(li);
+		}
+	}
+});
+
 
 ipcRenderer.on('update-view-toc', function(event, toc) {
 	//console.log("update toc");
@@ -26,14 +47,14 @@ ipcRenderer.on('update-view-toc', function(event, toc) {
 	let list = container.appendChild(document.createElement("ul"));
 	for (let section of toc) {
 		let li = document.createElement("li");
-		li.appendChild(linkOrText(section.title, section.url));
+		li.appendChild(linkOrText(section.title, {index: section.index, url: section.url}));
 		list.appendChild(li);
 
 		if ((section.subsections) && (section.subsections.length > 0)) {
 			let sublist = li.appendChild(document.createElement("ul"));
 			for (let subsection of section.subsections) {
 				let li2 = document.createElement("li");
-				li2.appendChild(linkOrText(subsection.title, subsection.url));
+				li2.appendChild(linkOrText(subsection.title, {index: subsection.index, url: subsection.url}));
 				sublist.appendChild(li2);
 			}
 		}
@@ -43,23 +64,39 @@ ipcRenderer.on('update-view-toc', function(event, toc) {
 });
 
 
-ipcRenderer.on('update-view-text', function(event, htmlText) {
-	//console.log("update text");
+ipcRenderer.on('update-view-text', function(event, sectionUrl, htmlText) {
+	//console.log("update-view-text");
+
 	let container = document.getElementById("textContainer");
 	container.innerHTML = htmlText;
+
+	let tocItems = document.getElementById("tocContainer").getElementsByTagName("a");
+	
+	for (let tocItem of tocItems) {
+		//console.log(tocItem.classList);
+		//console.log(tocItem.href); //file:///home/angian/git_workspace/angian-epub/Text/Section00xx.htm
+		//TODO: normalize urls to make more robust
+		if (tocItem.href.endsWith(sectionUrl)) {
+			tocItem.classList.add("current");
+		} else {
+			tocItem.classList.remove("current");
+		}
+	}
 });
 
 
-function linkOrText(text, targetUrl) {
+function linkOrText(text, section) {
 	let elem = null;
 
-	if (targetUrl && (targetUrl != "")) {
+	if (section) {
 		elem = document.createElement('a');  
 		let linkText = document.createTextNode(text);
-		elem.appendChild(linkText);  
-		//elem.href = targetUrl;
-		elem.href = "#";
-		elem.onclick = () => { ipcRenderer.send('update-model-text', targetUrl); };
+		elem.appendChild(linkText);
+		elem.href = section.url;
+		elem.onclick = () => {
+			ipcRenderer.send('update-model-text', section);
+			return false;
+		};
 
 	} else {
 		elem = document.createTextNode(text);