+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
 ## [Unreleased]
-
+### Added
+- navigate to next/previous section
+- navigate to bookmark
+- highlight current section in toc
 
 
 ## [0.5]